requirements.txt
torch>=2.0.0
torchvision>=0.15.0
mlflow>=2.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
grad-cam>=1.4.8
opencv-python>=4.8.0
python-multipart>=0.0.6
tqdm>=4.66.0
